// frontend/src/api/model/fieldResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */
import type { CoreType } from './coreType';
import type { FieldVisibility } from './fieldVisibility';
import type { Refinement } from './refinement';

export interface FieldResponse {
  name?: string;
  coreType?: CoreType;
  refinement?: Refinement;
  list?: boolean;
  required?: boolean;
  visibility?: FieldVisibility;
  description?: string;
}