// frontend/src/api/model/fieldVisibility.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */

export type FieldVisibility = typeof FieldVisibility[keyof typeof FieldVisibility];


export const FieldVisibility = {
  NORMAL: 'NORMAL',
  AUTO: 'AUTO',
  WRITE_ONLY: 'WRITE_ONLY',
} as const;